Initial commit of convolution example
Used the code written for an assignment in the course TDT4200 at NTNU as
a basis. Using `hipify-perl` to translate from cuda to hip did most of
the work. Some issues with using `extern "C"` and compiling with c++11,
so libs/bitmap.c is compiled with normal gcc and the linked.

The code seems to work, both I am not sure about the correctness of the
output. The resulting image after processing is not the same as when
running the serial version, or the parallel cuda version. Will
investigate that further.